refactor(ui): split guardrail_garden.tsx into focused modules
