+++ ui/litellm-dashboard/src/components/guardrails/guardrail_garden_card.tsx
@@ -1,0 +1,58 @@
+import React, { useState } from "react";
+import { CheckCircleFilled } from "@ant-design/icons";
+import { GuardrailCardInfo } from "./guardrail_garden_data";
+
+const GuardrailCard: React.FC<{ card: GuardrailCardInfo; onClick: () => void }> = ({ card, onClick }) => {
+  const [hovered, setHovered] = useState(false);
+
+  return (
+    <div
+      onClick={onClick}
+      onMouseEnter={() => setHovered(true)}
+      onMouseLeave={() => setHovered(false)}
+      style={{
+        borderRadius: 12,
+        border: hovered ? "1px solid #93c5fd" : "1px solid #e5e7eb",
+        backgroundColor: "#ffffff",
+        padding: "20px 20px 16px 20px",
+        cursor: "pointer",
+        transition: "border-color 0.15s, box-shadow 0.15s",
+        display: "flex",
+        flexDirection: "column",
+        minHeight: 170,
+        boxShadow: hovered ? "0 1px 6px rgba(59,130,246,0.08)" : "none",
+      }}
+    >
+      {/* Icon + Name row */}
+      <div style={{ display: "flex", alignItems: "center", gap: 10, marginBottom: 10 }}>
+        <img
+          src={card.logo}
+          alt=""
+          style={{ width: 28, height: 28, borderRadius: 6, objectFit: "contain", flexShrink: 0 }}
+          onError={(e) => { (e.target as HTMLImageElement).style.display = "none"; }}
+        />
+        <span style={{ fontSize: 14, fontWeight: 600, color: "#111827", lineHeight: 1.3 }}>{card.name}</span>
+      </div>
+
+      {/* Description */}
+      <p
+        className="line-clamp-3"
+        style={{ fontSize: 12, color: "#6b7280", lineHeight: 1.6, margin: 0, flex: 1 }}
+      >
+        {card.description}
+      </p>
+
+      {/* Eval badge */}
+      {card.eval && (
+        <div style={{ marginTop: 10, display: "flex", alignItems: "center", gap: 4 }}>
+          <CheckCircleFilled style={{ color: "#16a34a", fontSize: 12 }} />
+          <span style={{ fontSize: 11, color: "#16a34a", fontWeight: 500 }}>
+            F1: {card.eval.f1}% &middot; {card.eval.testCases} test cases
+          </span>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default GuardrailCard;
